爆発エフェクト進捗

+++ src/Entity/Effect/explosion.js
@@ -8,7 +8,10 @@ export default class Explosion extends EFFECT{
   constructor(name,pos,vel){
     if(name == "stone" || name == "smoke"){
       super(pos,vel);
-    }else{
+      //次のparticleを生成するかの変数
+      this.isNext = false;
+    }
+    else{
       super(pos,{x:0,y:0});
     }
     /*基本情報*/
@@ -23,6 +26,7 @@ export default class Explosion extends EFFECT{
     this.sprite = Art.SpriteFactory(this.pattern[this.spid]);
     //this.sprite = Art.SpriteFactory(this.pattern);
     this.sprite.position = this.pos;
+    if(this.name == "fire")this.sprite.alpha = 0.5;
   }
 
   Update(){
@@ -30,30 +34,46 @@ export default class Explosion extends EFFECT{
     this.sprite.anchor.set(0.5);
     switch(this.name){
       case "flash" :
-      if(this.frame == 4){
+      this.sprite.scale.x = 1;
+      this.sprite.scale.y = this.sprite.scale.x;
+      if(this.frame == 1){
         EntityManager.removeEntity(this);
       }
       break;
       case "fire" :
         let a = 10;
+        this.pos = Util.advec(this.pos,this.vel);
         this.sprite.scale.x += 1/(this.frame+2);
         this.sprite.scale.y = this.sprite.scale.x;
-        this.sprite.alpha = 1 - this.frame/10;
-      if(this.frame == 40){
+        this.sprite.alpha = 1 - this.frame/20;
+      if(this.frame == 10){
         EntityManager.removeEntity(this);
       }
       break;
       case "stone" :
-        //phys
-        this.pos.x += this.vel.x;
-        this.pos.y += this.vel.y;
-        this.sprite.scale.x *= 0.8;
-        this.sprite.scale.y = this.sprite.scale.x;
-        this.vel.y += 0.1;
+        //減速
+        this.vel.x *= 0.95;
+        this.vel.y *= 0.95;
+        this.pos.y += 0.3;//重力
+        //this.pos = Util.advec(this.pos,this.vel);
         this.sprite.alpha -= 0.02;
-      if(this.frame == 40){
-        EntityManager.removeEntity(this);
-      }
+        //再帰
+        if(this.sprite.alpha > 0 && this.isNext){
+          //生成は最初の一回のみ
+          this.isNext = false;
+          let p = Util.advec(this.pos,this.vel);
+          this.sprite.scale.x *= 0.8;
+          this.sprite.scale.y = this.sprite.scale.x;
+          let stone = new Explosion("stone",p,this.vel);
+          //次の石 : 小さく薄く
+          stone.sprite.scale.x = this.sprite.scale.x;
+          stone.sprite.scale.y = this.sprite.scale.y;
+          stone.sprite.alpha = this.sprite.alpha;
+          EntityManager.addEntity(stone);
+        }
+        if(this.frame == 1)this.isNext = true;
+        //持続時間
+        if(this.frame > 10)EntityManager.removeEntity(this);
       break;
       case "smoke" :
         let b = 10;
